Remove DataReferenceService changes

+++ src/main/java/bio/terra/workspace/service/datareference/flight/CreateDataReferenceStep.java
@@ -18,6 +18,9 @@ public class CreateDataReferenceStep implements Step {
 
   private DataReferenceDao dataReferenceDao;
 
+  private static final String CREATE_DATA_REFERENCE_COMPLETED_KEY =
+      "createDataReferenceStepCompleted";
+
   public CreateDataReferenceStep(DataReferenceDao dataReferenceDao) {
     this.dataReferenceDao = dataReferenceDao;
   }
@@ -26,7 +29,8 @@ public class CreateDataReferenceStep implements Step {
   public StepResult doStep(FlightContext flightContext) throws RetryException {
     FlightMap inputMap = flightContext.getInputParameters();
     FlightMap workingMap = flightContext.getWorkingMap();
-    UUID referenceId = workingMap.get(DataReferenceFlightMapKeys.REFERENCE_ID, UUID.class);
+    workingMap.put(CREATE_DATA_REFERENCE_COMPLETED_KEY, false);
+    UUID referenceId = inputMap.get(DataReferenceFlightMapKeys.REFERENCE_ID, UUID.class);
     UUID workspaceId = inputMap.get(DataReferenceFlightMapKeys.WORKSPACE_ID, UUID.class);
     DataRepoSnapshot reference =
         inputMap.get(DataReferenceFlightMapKeys.REFERENCE, DataRepoSnapshot.class);
@@ -42,8 +46,9 @@ public class CreateDataReferenceStep implements Step {
         body.getCloningInstructions(),
         body.getReferenceType(),
         reference);
+    workingMap.put(CREATE_DATA_REFERENCE_COMPLETED_KEY, true);
 
-    FlightUtils.setResponse(flightContext, referenceId, HttpStatus.OK);
+    FlightUtils.setResponse(flightContext, referenceId.toString(), HttpStatus.OK);
 
     return StepResult.getStepResultSuccess();
   }
@@ -53,11 +58,11 @@ public class CreateDataReferenceStep implements Step {
     FlightMap inputMap = flightContext.getInputParameters();
     FlightMap workingMap = flightContext.getWorkingMap();
 
-    UUID referenceId = workingMap.get(DataReferenceFlightMapKeys.REFERENCE_ID, UUID.class);
-    UUID workspaceId = inputMap.get(DataReferenceFlightMapKeys.WORKSPACE_ID, UUID.class);
-
-    // Ignore return value, as we don't care whether a reference was deleted or just not found.
-    dataReferenceDao.deleteDataReference(workspaceId, referenceId);
+    if (workingMap.get(CREATE_DATA_REFERENCE_COMPLETED_KEY, Boolean.class)) {
+      UUID referenceId = inputMap.get(DataReferenceFlightMapKeys.REFERENCE_ID, UUID.class);
+      UUID workspaceId = inputMap.get(DataReferenceFlightMapKeys.WORKSPACE_ID, UUID.class);
+      dataReferenceDao.deleteDataReference(workspaceId, referenceId);
+    }
 
     return StepResult.getStepResultSuccess();
   }
